get rid of treeFilter since there is no re-use
