+++ src/vs/workbench/contrib/debug/browser/media/repl.css
@@ -87,3 +87,40 @@
 .monaco-workbench .repl .repl-tree .output.expression .code-bold	{ font-weight: bold; }
 .monaco-workbench .repl .repl-tree .output.expression .code-italic	{ font-style: italic; }
 .monaco-workbench .repl .repl-tree .output.expression .code-underline { text-decoration: underline; }
+
+.monaco-action-bar .action-item.panel-action-tree-filter-container {
+	cursor: default;
+	display: flex;
+}
+
+.monaco-action-bar .panel-action-tree-filter{
+	display: flex;
+	align-items: center;
+	flex: 1;
+}
+
+.monaco-action-bar .panel-action-tree-filter .monaco-inputbox {
+	height: 24px;
+	font-size: 12px;
+	flex: 1;
+}
+
+.pane-header .monaco-action-bar .panel-action-tree-filter .monaco-inputbox {
+	height: 20px;
+	line-height: 18px;
+}
+
+.monaco-workbench.vs .monaco-action-bar .panel-action-tree-filter .monaco-inputbox {
+	height: 25px;
+}
+
+.panel > .title .monaco-action-bar .action-item.panel-action-tree-filter-container {
+	max-width: 400px;
+	min-width: 300px;
+	margin-right: 10px;
+}
+
+.monaco-action-bar .action-item.panel-action-tree-filter-container,
+.panel > .title .monaco-action-bar .action-item.panel-action-tree-filter-container.grow {
+	flex: 1;
+}
