--- src/vs/workbench/contrib/treeFilter/browser/media/treeFilter.css
@@ -1,41 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-.monaco-action-bar .action-item.panel-action-tree-filter-container {
-	cursor: default;
-	display: flex;
-}
-
-.monaco-action-bar .panel-action-tree-filter{
-	display: flex;
-	align-items: center;
-	flex: 1;
-}
-
-.monaco-action-bar .panel-action-tree-filter .monaco-inputbox {
-	height: 24px;
-	font-size: 12px;
-	flex: 1;
-}
-
-.pane-header .monaco-action-bar .panel-action-tree-filter .monaco-inputbox {
-	height: 20px;
-	line-height: 18px;
-}
-
-.monaco-workbench.vs .monaco-action-bar .panel-action-tree-filter .monaco-inputbox {
-	height: 25px;
-}
-
-.panel > .title .monaco-action-bar .action-item.panel-action-tree-filter-container {
-	max-width: 600px;
-	min-width: 300px;
-	margin-right: 10px;
-}
-
-.monaco-action-bar .action-item.panel-action-tree-filter-container,
-.panel > .title .monaco-action-bar .action-item.panel-action-tree-filter-container.grow {
-	flex: 1;
-}
